refactor: attempt at reducing bundle size
