+++ .github/workflows/publish.yml
@@ -1,7 +1,7 @@
 name: Publish
 on:
    workflow_run:
-      workflows: [CI]
+      workflows: [Lint and Type]
       branches: [main]
       types: [completed]
 
@@ -15,6 +15,9 @@ jobs:
    publish:
       if: ${{ github.event.workflow_run.conclusion == 'success' }}
       runs-on: ubuntu-latest
+      strategy:
+         matrix:
+            node-version: [18]
       steps:
          - name: Checkout code
            uses: actions/checkout@v3
@@ -27,11 +30,15 @@ jobs:
          - name: Setup Node.js
            uses: actions/setup-node@v3
            with:
-              node-version: 20.x
+              node-version: ${{ matrix.node-version }}
+              registry-url: "https://registry.npmjs.org"
               cache: "pnpm"
+              cache-dependency-path: "**/pnpm-lock.yaml"
 
-         - name: Install dependencies
-           run: pnpm install --frozen-lockfile
+         - name: Install dependencies and build
+           run: |
+              pnpm install --frozen-lockfile
+              pnpm build
 
          - name: Create Release Pull Request or Publish
            id: changesets
