+++ tsup.config.ts
@@ -2,13 +2,15 @@ import { defineConfig } from "tsup";
 
 export default defineConfig((options) => ({
 	format: ["esm", "cjs"],
-	entry: ["src/**/*.ts", "!src/types.ts"],
-	target: "node18",
+	entry: ["src/index.ts"],
+	target: "es2022",
 	clean: true, // clean up dist folder,
 	skipNodeModulesBundle: true, // skip building deps for node_modules, simply use them as is
 	dts: true,
+	splitting: false,
 	minify: !options.watch,
 	bundle: false,
 	treeshake: true,
+	tsconfig: "tsconfig.json",
 	sourcemap: !options.watch,
 }));
